+++ crates/egui_demo_lib/src/demo/plot_demo.rs
@@ -755,19 +755,38 @@ struct InteractionDemo {}
 impl InteractionDemo {
     #[allow(clippy::unused_self)]
     fn ui(&mut self, ui: &mut Ui) -> Response {
-        let plot = Plot::new("interaction_demo").height(300.0).legend(Legend::default());
+        let plot = Plot::new("interaction_demo")
+            .height(300.0)
+            .legend(Legend::default());
 
         let InnerResponse {
             response,
-            inner: (screen_pos, pointer_coordinate, pointer_coordinate_drag_delta, bounds, hovered,  hovered_indexes),
+            inner:
+                (
+                    screen_pos,
+                    pointer_coordinate,
+                    pointer_coordinate_drag_delta,
+                    bounds,
+                    hovered,
+                    hovered_indexes,
+                ),
         } = plot.show(ui, |plot_ui| {
-
-            plot_ui.line(  Line::new(PlotPoints::from_explicit_callback(
-                move |x| x.sin(),
-                ..,
-                100,
-                
-            )).name( "Sin"));
+            plot_ui.line(
+                Line::new(PlotPoints::from_explicit_callback(
+                    move |x| x.sin(),
+                    ..,
+                    100,
+                ))
+                .name("Sin"),
+            );
+            plot_ui.line(
+                Line::new(PlotPoints::from_explicit_callback(
+                    move |x| x.cos(),
+                    ..,
+                    110,
+                ))
+                .name("cos"),
+            );
             (
                 plot_ui.screen_from_plot(PlotPoint::new(0.0, 0.0)),
                 plot_ui.pointer_coordinate(),
@@ -802,10 +821,7 @@ impl InteractionDemo {
             "pointer coordinate drag delta: {}",
             coordinate_text
         ));
-        ui.label(format!(
-            "pointer is hovered_indexes: {:?}",
-            hovered_indexes
-        ));
+        ui.label(format!("pointer is hovered_indexes: {:?}", hovered_indexes));
 
         response
     }